Separate hit includes
