+++ include/hit/braceexpr.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "parse.h"
+#include "hit/parse.h"
 
 #include <string>
 #include <vector>
